pyproject.toml
[tool.poetry]
name = "lingopod"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohappyeyeballs = "^2.4.3"
aiohttp = "^3.10.10"
aiosignal = "^1.3.1"
alembic = "^1.14.0"
annotated-types = "^0.7.0"
anyio = "^4.6.2"
attrs = "^24.2.0"
bcrypt = "4.0.1"
beautifulsoup4 = "^4.12.3"
certifi = "^2024.8.30"
cffi = "^1.17.1"
charset-normalizer = "^3.4.0"
click = "^8.1.7"
colorama = "^0.4.6"
cryptography = "^43.0.3"
distro = "^1.9.0"
dnspython = "^2.7.0"
ecdsa = "^0.19.0"
edge-tts = "^6.1.17"
email-validator = "^2.2.0"
fastapi = "^0.115.4"
frozenlist = "^1.5.0"
greenlet = "^3.1.1"
h11 = "^0.14.0"
httpcore = "^1.0.6"
httpx = "^0.27.2"
idna = "^3.10"
iniconfig = "^2.0.0"
jiter = "^0.7.0"
jsonpatch = "^1.33"
jsonpointer = "^3.0"
langchain-core = "^0.3.15"
langchain-openai = "^0.2.6"
langsmith = "^0.1.142"
loguru = "^0.7.2"
mako = "^1.3.6"
markupsafe = "^3.0.2"
multidict = "^6.1.0"
openai = "^1.54.3"
orjson = "^3.10.11"
packaging = "^24.2"
passlib = "^1.7.4"
pluggy = "^1.5.0"
propcache = "^0.2.0"
pyasn1 = "^0.6.1"
pycparser = "^2.22"
pydantic = {version = "^2.9.2", extras = ["email"]}
pydantic-core = "^2.23.4"
pydantic-settings = "^2.6.1"
pydub = "^0.25.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
python-dotenv = "^1.0.1"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
python-multipart = "^0.0.17"
pyyaml = "^6.0.2"
regex = "^2024.11.6"
requests = "^2.32.3"
requests-toolbelt = "^1.0.0"
rsa = "^4.9"
six = "^1.16.0"
sniffio = "^1.3.1"
soupsieve = "^2.6"
sqlalchemy = "^2.0.36"
starlette = "^0.41.2"
tenacity = "^9.0.0"
tiktoken = "^0.8.0"
tqdm = "^4.67.0"
typing-extensions = "^4.12.2"
urllib3 = "^2.2.3"
uvicorn = "^0.32.0"
win32-setctime = "^1.1.0"
yarl = "^1.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
